MemorySegment being compatible with MemoryManagerOverMMF
A big one...

Introduced MMFRegistry to define an OS wide ID to each MMF.
MemorySegment now flags if the memory is from a regular MemoryManager or MMoMMF, for the later we don't store the address (which is process dependent) but MMFId and offset, the address is resolved at runtime.

UnmanagedList optimized and compatible with MMoMMF.

Fixed/added methods in BitMapHelpers with Unit Tests.