UnmanagedCollections and UnmanagedDataStore design and core implementation is done!
UnmanagedCollection are now implementing the IUnmanagedCollection interface. There is a new 3 layers principle:
 - The MemoryBlock storing the data, process-independent
 - The UnmanagedCollection which is the user facade, and process-dependent
 - The UnmanagedDataStore which stores the UnmanagedCollections and allow a ref access to the instances.
UnmanagedCollections can be stored in a MMF file (only the MemoryBlock is present), the store will have a special implementation for stored instances to work as expected.

The DefaultMemoryManager now exposes a Store and implement an IPageAllocator.

MMoMMF doesn't support a store yet, some data must be stored inside the MMF, other in process-dependent memory.

MemorySegment.Slice now allow negative offset to start from the end of the segment instead of the beginning.

Many little fixes, more documentation